Fix: Remove deprecated runtime "alpine"
The "alpine" runtime was not maintained and can be classified as
deprecated.

This deletes the repository, moving the init scripts to the
"aleph-debian-11-python" directory.
