--- runtimes/aleph-alpine-3.13-python/update_inits.sh
@@ -1,16 +1,0 @@
-#!/bin/sh
-
-umount /mnt/rootfs
-
-set -euf
-
-mount ./rootfs.ext4 /mnt/rootfs
-
-cp ./init0.sh /mnt/rootfs/sbin/init
-cp ./init1.py /mnt/rootfs/root/init1.py
-chmod +x /mnt/rootfs/sbin/init
-chmod +x /mnt/rootfs/root/init1.py
-
-umount /mnt/rootfs
-
-echo "OK"
